requirements.txt
aioredis==1.0.0
asgiref==2.3.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
channels==2.0.2
channels-redis==2.1.1
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
cryptography==2.2.1
daphne==2.1.0
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-jet==1.0.7
django-timezones==0.2
django-tinymce==2.7.0
django-widget-tweaks==1.3
gunicorn==19.7.1
h2==3.0.1
hiredis==0.2.0
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.6
oauthlib==2.0.7
Pillow==5.0.0
priority==1.3.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
service-identity==17.0.0
simplejson==3.13.2
six==1.11.0
South==1.0.2
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
websocket-client==0.47.0
zope.interface==4.4.3
